feat(cost-analysis): add price monitoring data and refactor related components
